pyproject.toml
[tool.poetry]
name = "db-ia-store"
version = "0.1.0"
description = ""
authors = ["jesustr20 <jesus.tr61@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = {extras = ["all"], version = "^0.115.6"}
sqlalchemy = "^2.0.36"
google-generativeai = "^0.8.3"
twilio = "^9.4.1"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
